+++ sites/public/pages/listings.tsx
@@ -99,6 +99,13 @@ const ListingsPage = () => {
     filterState
   )
 
+  const numberOfFilters = filterState
+    ? Object.keys(filterState).filter((p) => p !== "$comparison").length
+    : 0
+  const buttonTitle = numberOfFilters
+    ? t("listingFilters.buttonTitleWithNumber", { number: numberOfFilters })
+    : t("listingFilters.buttonTitle")
+
   const pageTitle = `${t("pageTitle.rent")} - ${t("nav.siteTitle")}`
   const metaDescription = t("pageDescription.welcome", { regionName: t("region.name") })
   const metaImage = "" // TODO: replace with hero image
@@ -195,8 +202,21 @@ const ListingsPage = () => {
           size={AppearanceSizeType.small}
           onClick={() => setFilterModalVisible(true)}
         >
-          {t("listingFilters.buttonTitle")}
+          {buttonTitle}
         </Button>
+        {numberOfFilters > 0 && (
+          <Button
+            className="mx-2 mt-6"
+            size={AppearanceSizeType.small}
+            styleType={AppearanceStyleType.secondary}
+            // "Submit" the form with no params to trigger a reset.
+            onClick={() => onSubmit({})}
+            icon="close"
+            iconPlacement="left"
+          >
+            {t("listingFilters.resetButton")}
+          </Button>
+        )}
       </div>
       {!listingsLoading && !listingsError && listingsData?.meta.totalItems === 0 && (
         <div className="container max-w-3xl my-4 px-4 content-start mx-auto">