+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file. The format 
 
 (_Note:_ it our intention to improve our release process going forward by using [Semantic Versioning](https://semver.org/spec/v2.0.0.html).)
 
+## Detroit Team M11
+
+- Added:
+
+  - Reset filter button
+    ([#489](https://github.com/CityOfDetroit/bloom/pull/489))
+  - Number of filters in filter button
+    ([#489](https://github.com/CityOfDetroit/bloom/pull/489))
+
 ## Detroit Team M10
 
 - Fixed: